Created lib/graphql Folder & Moved Client & Queries To The Folder
