+++ antt3.github.io/src/lib/graphql/queries.js
@@ -1,0 +1,30 @@
+import { gql } from 'graphql-request'
+
+export const authorsQuery = gql`
+    query GetAuthors {
+        authors {
+            name
+            intro
+            bio
+            slug
+            picture {
+                url
+            }
+        }
+    }
+`
+
+export const projectsQuery = gql`
+    query GetProjects {
+        projects {
+            name
+            slug
+            description
+            demo
+            sourceCode
+            image {
+                url
+            }
+        }
+    }
+`
